# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Cookie {
  id: number;
  name: string;
  slug: string;
  description: string;
  priceCents: number;
  stock: number;
  tags: string[];
  imageUrl: string;
  featured: boolean;
}

export interface DeliveryZone {
  id: number;
  suburb: string;
  postcode: string;
  standardFeeCents: number;
  expressFeeCents: number;
  pickupAvailable: boolean;
}

export interface DeliveryCheck {
  serviceable: boolean;
  /** @nullable */
  suburb: string | null;
  zone: DeliveryZone | null;
}

export interface OrderItemInput {
  cookieId: number;
  /** @minimum 1 */
  quantity: number;
}

export type OrderInputFulfillment = typeof OrderInputFulfillment[keyof typeof OrderInputFulfillment];


export const OrderInputFulfillment = {
  delivery: 'delivery',
  pickup: 'pickup',
} as const;

export type OrderInputSpeed = typeof OrderInputSpeed[keyof typeof OrderInputSpeed];


export const OrderInputSpeed = {
  standard: 'standard',
  express: 'express',
} as const;

/**
 * @nullable
 */
export type OrderInputBoxSize = typeof OrderInputBoxSize[keyof typeof OrderInputBoxSize] | null;


export const OrderInputBoxSize = {
  NUMBER_4: 4,
  NUMBER_6: 6,
  NUMBER_12: 12,
} as const;

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  email: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 1 */
  suburb: string;
  streetAddress?: string;
  fulfillment: OrderInputFulfillment;
  speed?: OrderInputSpeed;
  /** @nullable */
  boxSize?: OrderInputBoxSize;
  /** @maxLength 240 */
  giftNote?: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}

export interface OrderItem {
  cookieId: number;
  name: string;
  quantity: number;
  unitPriceCents: number;
}

export type OrderFulfillment = typeof OrderFulfillment[keyof typeof OrderFulfillment];


export const OrderFulfillment = {
  delivery: 'delivery',
  pickup: 'pickup',
} as const;

export type OrderSpeed = typeof OrderSpeed[keyof typeof OrderSpeed];


export const OrderSpeed = {
  standard: 'standard',
  express: 'express',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  received: 'received',
  baking: 'baking',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
} as const;

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  email: string;
  phone: string;
  suburb: string;
  fulfillment: OrderFulfillment;
  speed: OrderSpeed;
  status: OrderStatus;
  items: OrderItem[];
  subtotalCents: number;
  gstCents: number;
  deliveryFeeCents: number;
  totalCents: number;
  giftNote: string;
  createdAt: string;
}

export interface Error {
  error: string;
}

export type CheckDeliveryZoneParams = {
/**
 * @minLength 1
 */
suburb: string;
};