# lib/api-zod/src/generated/types/cookie.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Cookie {
  id: number;
  name: string;
  slug: string;
  description: string;
  priceCents: number;
  stock: number;
  tags: string[];
  imageUrl: string;
  featured: boolean;
}