# lib/api-zod/src/generated/types/deliveryCheck.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryZone } from './deliveryZone';

export interface DeliveryCheck {
  serviceable: boolean;
  /** @nullable */
  suburb: string | null;
  zone: DeliveryZone | null;
}