# lib/api-zod/src/generated/types/deliveryZone.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeliveryZone {
  id: number;
  suburb: string;
  postcode: string;
  standardFeeCents: number;
  expressFeeCents: number;
  pickupAvailable: boolean;
}