# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderFulfillment } from './orderFulfillment';
import type { OrderItem } from './orderItem';
import type { OrderSpeed } from './orderSpeed';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNumber: string;
  customerName: string;
  email: string;
  phone: string;
  suburb: string;
  fulfillment: OrderFulfillment;
  speed: OrderSpeed;
  status: OrderStatus;
  items: OrderItem[];
  subtotalCents: number;
  gstCents: number;
  deliveryFeeCents: number;
  totalCents: number;
  giftNote: string;
  createdAt: Date;
}