# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputBoxSize } from './orderInputBoxSize';
import type { OrderInputFulfillment } from './orderInputFulfillment';
import type { OrderInputSpeed } from './orderInputSpeed';
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  email: string;
  /** @minLength 5 */
  phone: string;
  /** @minLength 1 */
  suburb: string;
  streetAddress?: string;
  fulfillment: OrderInputFulfillment;
  speed?: OrderInputSpeed;
  /** @nullable */
  boxSize?: OrderInputBoxSize;
  /** @maxLength 240 */
  giftNote?: string;
  /** @minItems 1 */
  items: OrderItemInput[];
}