# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  cookieId: number;
  name: string;
  quantity: number;
  unitPriceCents: number;
}